Hold the liveness cycle when the multiplexer does not answer
`batch_liveness_check_with_identity` returned an empty vector on probe
failure, and the caller read that as "no agents are dead". The two cases are
not the same, and collapsing them means a failed subprocess call quietly
asserts that everything is healthy.

It returns `Observed` now, so the caller has to choose. The choice is made
once, in `answered_dead_identities`, which logs the boundary and the reason
rather than dropping them -- an invisible hold looks identical to the
fail-open bug from the outside, and the operator needs to be able to tell.

Reconciliation is split into a pure `reconcile_observed` that takes both probe
results as arguments. That is what makes each boundary drivable to failure in
a test: no environment mutation, which is unavailable here anyway because
`set_var` is unsafe under edition 2024 and unsafe is forbidden.

If either probe is unknown the whole reconciliation is unknown. A session set
that is missing because `list-sessions` failed cannot be told apart from one
missing because the sessions ended, and guessing between those is how #527
marked twenty live panes stopped.

The fourth test guards the mirror hazard: with both probes answering, an
absent session is still reported dead. Fail-closed must not become
never-closed, and a change that only ever holds would satisfy the invariant
while breaking the feature.

Verified RED: restoring the old "unanswered means no sessions exist" shape
fails `an_unknown_session_list_produces_no_transition` on "a failed probe must
not decide".

The test module moved to `liveness_tests.rs` because the additions pushed
`liveness.rs` past the 1000-line gate; this follows the existing
`multiplexer_tests.rs` pattern. Two helpers widened to `pub(super)` to stay
reachable. No test content changed.

Refs #541
